pro/VBAIgame-main/vbaigame-speech/requirements_enhanced.txt
certifi==2024.8.30
packaging==24.1
pygame==2.6.1
openai==1.12.0
python-dotenv==1.0.0
PyOpenGL==3.1.6
numpy>=1.24.1
sounddevice==0.4.6
websockets==12.0
asyncio-mqtt==0.16.1
